remove cache
