+++ lib/core.js
@@ -1,115 +1,18 @@
 var path = require('path')
-var cheerio = require('cheerio')
-var hljs = require('highlight.js')
 var loaderUtils = require('loader-utils')
-var markdown = require('markdown-it')
-var cache = require('./cache')
-var genId = require('./gen-id')
 
-/**
- * `{{ }}` => `<span>{{</span> <span>}}</span>`
- * @param  {string} str
- * @return {string}
- */
-var replaceDelimiters = function (str) {
-  return str.replace(/({{|}})/g, '<span>$1</span>')
-}
-
-/**
- * renderHighlight
- * @param  {string} str
- * @param  {string} lang
- */
-var renderHighlight = function (str, lang) {
-  if (!(lang && hljs.getLanguage(lang))) {
-    return ''
-  }
-
-  try {
-    return replaceDelimiters(hljs.highlight(lang, str, true).value)
-  } catch (err) {}
-}
-
-/**
- * html => vue file template
- * @param  {[type]} html [description]
- * @return {[type]}      [description]
- */
-var renderVueTemplate = function (html) {
-  var $ = cheerio.load(html, {
-    decodeEntities: false,
-    lowerCaseAttributeNames: false,
-    lowerCaseTags: false
-  })
-
-  var output = {
-    style: $.html('style'),
-    script: $.html('script')
-  }
-  var result
-
-  $('style').remove()
-  $('script').remove()
-
-  result = '<template><section>' + $.html() + '</section></template>\n' +
-    output.style + '\n' +
-    output.script
-
-  return result
-}
+var markdownCompilerPath = path.resolve(__dirname, 'markdown-compiler.js')
 
 module.exports = function (source) {
   this.cacheable()
 
-  var parser
-  var params = loaderUtils.parseQuery(this.query)
-  var opts = Object.assign(params, this.vueMarkdown, this.options.vueMarkdown)
-
-  if ({}.toString.call(opts.render) === '[object Function]') {
-    parser = opts
-  } else {
-    opts = Object.assign({
-      preset: 'default',
-      html: true,
-      highlight: renderHighlight
-    }, opts)
-
-    var plugins = opts.use
-    var preprocess = opts.preprocess
-
-    delete opts.use
-    delete opts.preprocess
-
-    parser = markdown(opts.preset, opts)
-    if (plugins) {
-      plugins.forEach(function (plugin) {
-        if (Array.isArray(plugin)) {
-          parser.use.apply(parser, plugin)
-        } else {
-          parser.use(plugin)
-        }
-      })
-    }
-  }
-
-  var codeInlineRender = parser.renderer.rules.code_inline;
-  parser.renderer.rules.code_inline = function () {
-    return replaceDelimiters(codeInlineRender.apply(this, arguments));
-  }
-
-  if (preprocess) {
-    source = preprocess.call(this, parser, source)
-  }
-  source = source.replace(/@/g, '__at__')
+  this.options.__vueMarkdownOptions__ = this.query || {}
 
   var filePath = this.resourcePath
-  var content = parser.render(source).replace(/__at__/g, '@')
-  var result = renderVueTemplate(content)
-  var fileName = path.basename(filePath, '.md')
 
-  filePath = cache.save(fileName + '-' + genId(filePath), result)
+  var result = 'module.exports = require(' +
+    loaderUtils.stringifyRequest(this, '!!vue-loader!' + markdownCompilerPath + '?raw!' + filePath) +
+  ');'
 
-  return 'module.exports = require(' +
-    loaderUtils.stringifyRequest(this, '!!vue-loader!' + filePath) +
-    ');'
+  return result
 }
