+++ lib/markdown-compiler.js
@@ -1,0 +1,102 @@
+var loaderUtils = require('loader-utils')
+var cheerio = require('cheerio')
+var markdown = require('markdown-it')
+
+/**
+ * `{{ }}` => `<span>{{</span> <span>}}</span>`
+ * @param  {string} str
+ * @return {string}
+ */
+var replaceDelimiters = function (str) {
+  return str.replace(/({{|}})/g, '<span>$1</span>')
+}
+
+/**
+ * html => vue file template
+ * @param  {[type]} html [description]
+ * @return {[type]}      [description]
+ */
+var renderVueTemplate = function (html) {
+  var $ = cheerio.load(html, {
+    decodeEntities: false,
+    lowerCaseAttributeNames: false,
+    lowerCaseTags: false
+  })
+
+  var output = {
+    style: $.html('style'),
+    script: $.html('script')
+  }
+  var result
+
+  $('style').remove()
+  $('script').remove()
+
+  result = '<template><section>' + $.html() + '</section></template>\n' +
+    output.style + '\n' +
+    output.script
+
+  return result
+}
+
+module.exports = function (source) {
+  this.cacheable && this.cacheable()
+  var callback = this.async()
+  var parser
+  var params = loaderUtils.parseQuery(this.query) || {}
+  var vueMarkdownOptions = Object.create(this.options.__vueMarkdownOptions__ ? this.options.__vueMarkdownOptions__.__proto__ : {})
+  var opts = Object.assign(vueMarkdownOptions, params, this.options.__vueMarkdownOptions__, this.options.vueMarkdown)
+
+  if ({}.toString.call(opts.render) === '[object Function]') {
+    parser = opts
+  } else {
+    opts = Object.assign({
+      preset: 'default',
+      html: true,
+    }, opts)
+
+    var plugins = opts.use
+    var preprocess = opts.preprocess
+
+    delete opts.use
+    delete opts.preprocess
+
+    parser = markdown(opts.preset, opts)
+    if (plugins) {
+      plugins.forEach(function (plugin) {
+        if (Array.isArray(plugin)) {
+          parser.use.apply(parser, plugin)
+        } else {
+          parser.use(plugin)
+        }
+      })
+    }
+  }
+
+  var codeInlineRender = parser.renderer.rules.code_inline;
+  parser.renderer.rules.code_inline = function () {
+    return replaceDelimiters(codeInlineRender.apply(this, arguments));
+  }
+
+  if (preprocess) {
+    source = preprocess.call(this, parser, source)
+  }
+
+  function exportContent (content) {
+    if (opts.raw) {
+      callback(null, content)
+    } else {
+      callback(null, 'module.exports = ' + JSON.stringify(content))
+    }
+  }
+
+  source = source.replace(/@/g, '__at__')
+
+  var content = parser.render(source).replace(/__at__/g, '@')
+  var result = renderVueTemplate(content)
+
+  // for relative includes
+  opts.filename = this.resourcePath
+
+  exportContent(result)
+}
